% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1-create-hw-vars.R
\name{create_heatwave_vars}
\alias{create_heatwave_vars}
\title{Create heatwave variables}
\usage{
create_heatwave_vars(
  df,
  threshold_vars,
  consec_days,
  date_var,
  clim_var,
  psu_var
)
}
\arguments{
\item{df}{A data.table object containing the dataset.}

\item{threshold_vars}{A character vector containing the names of the threshold variables.}

\item{consec_days}{A numeric vector containing the number of consecutive days to define a heatwave.}

\item{date_var}{A character string specifying the name of the date variable.}

\item{clim_var}{A character string specifying the name of the climate variable.}

\item{psu_var}{A character string specifying the name of the PSU variable.}
}
\value{
A data.table object with heatwave variables.
}
\description{
This function creates heatwave variables based on the specified threshold values and consecutive days.
Note that threshold values need to be passed as variables, which makes this function more flexible to be used with absolute as well as relative threshold values.
}
