% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-lt-cutoff-vars-crude-and-rolling.R
\name{flexi_percentile_cutoffs}
\alias{flexi_percentile_cutoffs}
\title{Calculate the rolling percentile for each date and geographic unit (e.g. Zip/psu)}
\usage{
flexi_percentile_cutoffs(
  DT,
  var_col = "tmax",
  ntile = 0.9,
  perc_type = "crude",
  num_days = 3,
  psu_col = "Zip",
  num_cores_leave = 1
)
}
\arguments{
\item{DT}{A data.table or dataframe with columns 'date', variable of interest (e.g. tmax), and the geographic unit (e.g. Zip/PSU)}

\item{var_col}{The name of the column containing the variable of interest}

\item{ntile}{The percentile to calculate (default is 0.9)}

\item{perc_type}{The type of percentile to calculate (default is 'crude'). Can alse be "rolling". If 'rolling', num_days must be specified. Can also be "crude_doy"}

\item{num_days}{The number of days to consider in the rolling window (default is 3)}

\item{psu_col}{The name of the column containing the geographic unit (default is 'Zip')}

\item{num_cores_leave}{The number of cores to leave for other processes (default is 1)}
}
\value{
The input data.table with an additional column 'percentile' containing the rolling/crude percentile
}
\description{
Calculate the rolling percentile for each date and geographic unit (e.g. Zip/psu)
}
