% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-variables-between-a-period.R
\name{func_calc_bins_period}
\alias{func_calc_bins_period}
\title{Calculate the number of instances for each bin of a climatic variable within a period}
\usage{
func_calc_bins_period(
  df_health,
  df_lt_clim,
  start_date_var,
  add_days,
  subtract_days,
  psu_var = "psu",
  clim_var = "max_temp_wb"
)
}
\arguments{
\item{df_health}{A data.table containing the health data}

\item{df_lt_clim}{A data.table containing the long-term climatic data}

\item{start_date_var}{A character string specifying the column name of the start date}

\item{add_days}{An integer specifying the number of days to add to the start date}

\item{subtract_days}{An integer specifying the number of days to subtract from the start date}

\item{psu_var}{A character string specifying the column name of the PSU}

\item{clim_var}{A character string specifying the column name of the climatic variable}
}
\value{
A data.table containing the health data with the number of instances for each bin
}
\description{
This function calculates the number of instances for each bin of a climatic variable within a period.
It takes as input a regular time series dataset with a climatic variable and a health dataset with a start date and calculates the number of instances for each bin of the climatic variable within a period.
}
