% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.1-create-variables-between-a-period.R
\name{func_calc_descr_period}
\alias{func_calc_descr_period}
\title{Calculate various descriptive statistics for a climatic variable from a time series dataframe within a period for a corresponding health data}
\usage{
func_calc_descr_period(
  df_lt_clim,
  df_health,
  start_date_var,
  add_days,
  subtract_days,
  psu_var,
  sum = TRUE,
  count_non_zero = TRUE,
  max = TRUE,
  count_d = FALSE,
  d = NULL,
  vec_identifiers
)
}
\arguments{
\item{df_lt_clim}{A data.table containing the long-term climatic data}

\item{df_health}{A data.table containing the health data}

\item{start_date_var}{A character string specifying the column name of the start date}

\item{add_days}{An integer specifying the number of days to add to the start date}

\item{subtract_days}{An integer specifying the number of days to subtract from the start date}

\item{psu_var}{A character string specifying the column name of the PSU}

\item{sum}{A logical indicating whether to calculate the sum (default: TRUE), Can be set to FALSE to skip the calculation}

\item{count_non_zero}{A logical indicating whether to calculate the count of non-zero values (default: TRUE), Can be set to FALSE to skip the calculation}

\item{max}{A logical indicating whether to calculate the max (default: TRUE), Can be set to FALSE to skip the calculation}

\item{count_d}{A logical indicating whether to calculate the count of a specific value (default: FALSE), Can be set to TRUE to calculate the count of a specific value}

\item{d}{A numeric value specifying the value to count. Must only be provided if count_d is TRUE.}

\item{vec_identifiers}{A character vector containing the identifiers of the columns to calculate the sum, count, and max}
}
\description{
This function calculates various descriptive statistics for a climatic variable from a time series dataframe within a period for a corresponding health data.
Note: the psu vars need to have the same name in both data.tables and both need to be the same data type.
}
\examples{
 \dontrun{
vec_identifiers <- c("consec")
df_test <- func_calc_descr_period(
    df_lt_clim = df_lt_vars_2014, 
    df_health = df_anc_3mo, 
    start_date_var = "dob", 
    add_days = NULL,
    subtract_days = 7, 
    psu_var = "psu",
    vec_identifiers = vec_identifiers,
    sum = TRUE,
    count_non_zero = TRUE,
    max = FALSE
    cound_d = TRUE,
    d = 5)
}
}
