% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-variables-between-a-period.R
\name{func_calc_sum_count_exceed_period}
\alias{func_calc_sum_count_exceed_period}
\title{Calculate the sum and count of exceedances for each individual for a specific period}
\usage{
func_calc_sum_count_exceed_period(
  df_lt_proc_exceed,
  df_health,
  start_date_var,
  add_days,
  subtract_days,
  psu_var
)
}
\arguments{
\item{df_lt_proc_exceed}{A data.table object with columns: date, tmax, extreme_event, and exceedance.}

\item{df_health}{A data.table object with columns: start_date_var, psu_var, add_days, and subtract_days.}

\item{start_date_var}{The name of the start date variable.}

\item{add_days}{The number of days to add to the start date. Cant be NULL if subtract_days is NULL.}

\item{subtract_days}{The number of days to subtract from the start date. Cant be NULL if add_days is NULL.}

\item{psu_var}{The name of the PSU variable.}
}
\value{
A data.table object with sum and count variables for exceedances.
}
\description{
This function calculates the sum and count of exceedances for each individual for a specific period.
It takes the outputs of the func_ident_mag_calc_mag function and calculates the sum and count of exceedances for each individual for a specific period.
}
