% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2-extract-clim-data-for-points.R
\name{func_extract_clim_data_point}
\alias{func_extract_clim_data_point}
\title{Extract climate data for a point shape file
The function reads the NetCDF files, extracts the climate data for each point of interest, and compiles the mean climate for each day.
The function assumes that the climate data is daily and works only for point shape files like DHS data.}
\usage{
func_extract_clim_data_point(
  path,
  region_boundary = india_boundary_buf,
  dhs_lat_long_sf = df_dhs_psu_geo_sf,
  clim_var = "tmax_wb",
  start_index = 1,
  end_index = NULL
)
}
\arguments{
\item{path}{Path to the NetCDF files}

\item{region_boundary}{Region boundary for the climate data}

\item{dhs_lat_long_sf}{Latitude and longitude of the DHS data}

\item{clim_var}{Climate variable to extract}

\item{start_index}{Start index of the NetCDF files}

\item{end_index}{End index of the NetCDF files}
}
\value{
A data frame with the extracted climate data
}
\description{
Extract climate data for a point shape file
The function reads the NetCDF files, extracts the climate data for each point of interest, and compiles the mean climate for each day.
The function assumes that the climate data is daily and works only for point shape files like DHS data.
}
