% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-clim-data-gridded-nics.R
\name{func_extract_clim_data_shp}
\alias{func_extract_clim_data_shp}
\title{This function extracts climate data from a nic file for a given shapefile and attribute}
\usage{
func_extract_clim_data_shp(
  path_nic_files,
  shape_file,
  admin_colname = "full_region",
  var_name = "tmax_noaa"
)
}
\arguments{
\item{path_nic_files}{The path to the folder containing the NIC files}

\item{admin_colname}{The name of the attribute in the shapefile that identifies the region
for which the climate variable needs to be extracted (e.g. "Zip"). Use "full_region" if the data needs to be extracted for the entire region.}

\item{sf_file}{The shapefile containing the areas of interest. Currently handles sf or SpatVect files.}
}
\description{
This function extracts climate data from a nic file for a given shapefile and attribute
}
