% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2-extract-clim-data-gridded-nics-no-date.R
\name{func_extract_clim_data_shp_nodate}
\alias{func_extract_clim_data_shp_nodate}
\title{Extract climate data for a shape file when the NetCDF files do not have date information}
\usage{
func_extract_clim_data_shp_nodate(
  path_nic_files,
  sf_file,
  sf_file_admin,
  reference_date = "1900-01-01",
  file_frequency = "daily"
)
}
\arguments{
\item{path_nic_files}{Path to the NetCDF files}

\item{sf_file}{Shape file}

\item{sf_file_admin}{Administrative variable in the shape file}

\item{reference_date}{A character string specifying the reference date for the NetCDF files}

\item{file_frequency}{A character string specifying the frequency of the NetCDF files}
}
\value{
A data frame with the extracted climate data
}
\description{
This function extracts climate data for a shape file, when the NetCDF files do not have date information.
The function reads the NetCDF files, extracts the climate data for each area of interest, and compiles the mean climate for each day.
The function assumes that the climate data is daily and works only for regions like Zip codes, states, or countries.
}
