% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-clim-data-gridded-nics.R
\name{func_extract_dates_nic}
\alias{func_extract_dates_nic}
\title{Extract dates from NIC files}
\usage{
func_extract_dates_nic(path_nic_files)
}
\arguments{
\item{path_nic_files}{The path to the folder containing the NIC files}
}
\value{
A vector of dates extracted from the NIC files
}
\description{
This function extracts dates from the nic files
}
