% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-sum-and-count-vars-for-exceedances.R
\name{function_calc_sum_count_exceed}
\alias{function_calc_sum_count_exceed}
\title{Calculate the sum and count of exceedances for each individual}
\usage{
function_calc_sum_count_exceed(
  df_input,
  df_target,
  start_date_var,
  add_days,
  subtract_days,
  psu_var
)
}
\description{
Calculate the sum and count of exceedances for each individual
}
