% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-heat-index-and-apparent-temp.R
\name{heat_index}
\alias{heat_index}
\title{Calculate heat index and apparent temperature}
\usage{
heat_index(temperature, rh, dewpoint, dp, t_f, dp_f, rh_original)
}
\arguments{
\item{temperature}{Air temperature in Fahrenheit or Celsius}

\item{rh}{Relative humidity in percentage (0-100) or NA if using dew point}

\item{dewpoint}{Dew point temperature in Fahrenheit or Celsius or NA if using relative humidity}

\item{dp}{TRUE or FALSE; TRUE if dew point temperature is used}

\item{t_f}{TRUE or FALSE; TRUE if air temperature is provided in Fahrenheit}

\item{dp_f}{TRUE or FALSE; TRUE if dew point temperature is provided in Fahrenheit}

\item{rh_original}{TRUE or FALSE; TRUE if relative humidity ranges from 0 to 1 and will be transformed to a range of 0 to 100}
}
\value{
A data.frame of heat index in Fahrenheit/Celsius and apparent temperature in Fahrenheit/Celsius
}
\description{
Calculate heat index and apparent temperature
}
