% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatre-TED-event-vars.R
\name{ted_events}
\alias{ted_events}
\title{Calculate temperature exceedance duration events}
\usage{
ted_events(
  df,
  var_raw = "tmax",
  var_cutoff = "extreme_event",
  var_exceed = "exceedance"
)
}
\arguments{
\item{df}{A data.table object with columns: date, tmax, extreme_event, and exceedance.}

\item{var_raw}{The name of the raw variable (default: "tmax").}

\item{var_cutoff}{The name of the variable indicating the extreme event (default: "extreme_event").}

\item{var_exceed}{The name of the variable indicating the exceedance (default: "exceedance").}
}
\value{
A data.table object with summarized extreme events.
}
\description{
Calculate temperature exceedance duration events
}
