% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-sum-and-count-vars-for-exceedances.R
\name{func_calc_exceed_sum_count}
\alias{func_calc_exceed_sum_count}
\alias{func_ident_mag_calc_mag}
\title{Calculate the sum and count of exceedances for each individual}
\usage{
func_ident_mag_calc_mag(df, vec_threshold_abs, vec_varnames_perc, clim_var)
}
\arguments{
\item{df}{A data.table object with columns: date, tmax, extreme_event, and exceedance.}

\item{vec_threshold_abs}{A vector of absolute thresholds. Can also be NULL.}

\item{vec_varnames_perc}{A vector of percentile thresholds. Can also be NULL.}

\item{clim_var}{The name of the climate variable.}
}
\value{
A data.table object with sum and count variables for exceedances.
}
\description{
This function takes a time series and identifies days that exceed a threshold. It also calculates the magnitude of exceedances for each day.
}
